backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-otp==1.2.2
qrcode[pil]==7.4.2